backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
scikit-image
ncempy
matplotlib
pandas
pytesseract
Pillow
stardist
tensorflow
openpyxl
